% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_ms_file_descriptions}
\alias{get_ms_file_descriptions}
\title{get_ms_file_descriptions}
\usage{
get_ms_file_descriptions()
}
\value{
A \linkS4class{data.table} with details for each ".mzML" and
".mzXML" file.
}
\description{
A table with a description of the ".mzML" and ".mzXML" files
in the \pkg{StreamFindData} package.
}
