% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_ms_mrm_spiked_estrogens}
\alias{get_ms_mrm_spiked_estrogens}
\title{get_ms_mrm_spiked_estrogens}
\usage{
get_ms_mrm_spiked_estrogens()
}
\value{
A \linkS4class{data.table} with the list of estrogens spiked in
sample corresponding to the ".mzML" file in MRM mode with negative polarity.
}
\description{
List of estrogens spiked to ".mzML" files in MRM mode.
}
