% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_ms_tof_spiked_chemicals_with_ms2}
\alias{get_ms_tof_spiked_chemicals_with_ms2}
\title{get_ms_tof_spiked_chemicals_with_ms2}
\usage{
get_ms_tof_spiked_chemicals_with_ms2()
}
\value{
A \linkS4class{data.table} with the list of chemicals spiked in
samples corresponding to the TOF ".mzML" and ".mzXML" files.
}
\description{
List of chemicals spiked to TOF ".mzML" and ".mzXML" files
including fragmentation pattern acquired with DDA.
}
