% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{get_raman_file_paths}
\alias{get_raman_file_paths}
\title{get_raman_file_paths}
\usage{
get_raman_file_paths()
}
\value{
A character vector with the full paths of Raman files.
}
\description{
Gets the full paths of ".asc" Raman files in the
\pkg{StreamFindData} package.
}
