% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_tof_spiked_chemicals}
\alias{get_tof_spiked_chemicals}
\title{get_tof_spiked_chemicals}
\usage{
get_tof_spiked_chemicals()
}
\value{
A \linkS4class{data.table} with the list of chemicals spiked in
samples corresponding to the ".mzML" and ".mzXML" files.
}
\description{
List of chemicals spiked to ".mzML" and ".mzXML" files.
}
