% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{msFilePaths}
\alias{msFilePaths}
\title{msFilePaths}
\usage{
msFilePaths()
}
\value{
A character vector with the full paths of MS files.
}
\description{
Gets the full paths of ".mzML" and ".mzXML" files in the
\pkg{streamFindData} package.
}
