% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{msFilesDescription}
\alias{msFilesDescription}
\title{msFilesDescription}
\usage{
msFilesDescription()
}
\value{
A \linkS4class{data.table} with details for each ".mzML" and
".mzXML" file.
}
\description{
A table with a description of the ".mzML" and ".mzXML" files
in the \pkg{streamFindData} package.
}
