% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{msSpikedChemicals}
\alias{msSpikedChemicals}
\title{msSpikedChemicals}
\usage{
msSpikedChemicals()
}
\value{
A \linkS4class{data.table} with the list of chemicals spiked in
samples corresponding to the ".mzML" and ".mzXML" files.
}
\description{
List of chemicals spiked to ".mzML" and ".mzXML" files.
}
