% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{msSpikedEstrogens}
\alias{msSpikedEstrogens}
\title{msSpikedEstrogens}
\usage{
msSpikedEstrogens()
}
\value{
A \linkS4class{data.table} with the list of estrogens spiked in
sample corresponding to the ".mzML" file in MRM mode with negative polarity.
}
\description{
List of estrogens spiked to ".mzML" files in MRM mode.
}
