% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamFindData-package.R
\docType{package}
\name{streamFindData-package}
\alias{streamFindData}
\alias{streamFindData-package}
\title{streamFindData: Package to hold supporting data for package streamFind.}
\description{
Package to hold supporting data for package streamFind..
}
\keyword{internal}
